+++ db/migrate/20200218144849_drop_material_parts_table.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class DropMaterialPartsTable < Lcms::Engine::Migration
+class DropMaterialPartsTable < ActiveRecord::Migration[5.2]
   def change
     drop_table :material_parts
   end