+++ db/migrate/20200422130652_add_reimported_at_to_documents.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class AddReimportedAtToDocuments < Lcms::Engine::Migration
+class AddReimportedAtToDocuments < ActiveRecord::Migration[5.2]
   def change
     add_column :documents, :reimported_at, :datetime
   end