+++ db/migrate/20200429130353_drop_pages.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class DropPages < Lcms::Engine::Migration
+class DropPages < ActiveRecord::Migration[5.2]
   def change
     drop_table :pages do |t|
       t.text :body, null: false