+++ lib/doc_template.rb
@@ -60,12 +60,12 @@ module DocTemplate
   end
 end
 
-require_dependency 'doc_template/template'
-require_dependency 'doc_template/document'
-require_dependency 'doc_template/tags'
-require_dependency 'doc_template/document_toc'
-require_dependency 'doc_template/xpath_functions'
+require 'doc_template/template'
+require 'doc_template/document'
+require 'doc_template/tags'
+require 'doc_template/document_toc'
+require 'doc_template/xpath_functions'
 
-Dir["#{__dir__}/doc_template/tables/*.rb"].each(&method(:require_dependency))
-Dir["#{__dir__}/doc_template/tags/*.rb"].each(&method(:require_dependency))
-Dir["#{__dir__}/doc_template/objects/*.rb"].each(&method(:require_dependency))
+Dir["#{__dir__}/doc_template/tables/*.rb"].sort.each(&method(:require))
+Dir["#{__dir__}/doc_template/tags/*.rb"].sort.each(&method(:require))
+Dir["#{__dir__}/doc_template/objects/*.rb"].sort.each(&method(:require))